+++ components/screens/ActiveVideoRoom/Participant/EmojiReactionOverlay/EmojiReactionOverlay.tsx
@@ -1,0 +1,30 @@
+import { RemoteTrackPublication, LocalTrackPublication } from "twilio-video";
+import useRemoteEmojiDataTrack from "../../../../../lib/hooks/useRemoteEmojiDataTrack";
+import { useVideoStore, VideoAppState } from "../../../../../store/store";
+
+import { OverlayEmoji } from "../../../../styled";
+
+interface EmojiReactionOverlayProps {
+  isLocalParticipant: boolean;
+  emojiPublication: LocalTrackPublication | RemoteTrackPublication | undefined;
+}
+
+export default function EmojiReactionOverlay({
+  isLocalParticipant,
+  emojiPublication,
+}: EmojiReactionOverlayProps) {
+  const { localEmoji } = useVideoStore((state: VideoAppState) => state);
+  const remoteEmoji = !isLocalParticipant
+    ? useRemoteEmojiDataTrack(emojiPublication)
+    : null;
+
+  return (
+    <OverlayEmoji>
+      {isLocalParticipant && localEmoji ? (
+        <>{localEmoji}</>
+      ) : (
+        <>{remoteEmoji && <>{remoteEmoji}</>}</>
+      )}
+    </OverlayEmoji>
+  );
+}
